Change argument ms of function st_timer_sleep to unsigned

+++ include/steroids/timer.h
@@ -9,7 +9,7 @@ static void st_timer_quit(st_modctx_t *timer_ctx);
 
 static uint64_t st_timer_start(st_modctx_t *timer_ctx);
 static unsigned st_timer_get_elapsed(st_modctx_t *timer_ctx, uint64_t start);
-static void st_timer_sleep(st_modctx_t *timer_ctx, uint64_t ms);
+static void st_timer_sleep(st_modctx_t *timer_ctx, unsigned ms);
 static void st_timer_sleep_for_fps(st_modctx_t *timer_ctx, unsigned fps);
 
 #endif /* ST_STEROIDS_TIMER_H */
