+++ include/steroids/types/modules/timer.h
@@ -9,7 +9,7 @@ typedef void (*st_timer_quit_t)(st_modctx_t *timer_ctx);
 typedef uint64_t (*st_timer_start_t)(st_modctx_t *timer_ctx);
 typedef unsigned (*st_timer_get_elapsed_t)(st_modctx_t *timer_ctx,
  uint64_t start);
-typedef void (*st_timer_sleep_t)(st_modctx_t *timer_ctx, uint64_t ms);
+typedef void (*st_timer_sleep_t)(st_modctx_t *timer_ctx, unsigned ms);
 typedef void (*st_timer_sleep_for_fps_t)(st_modctx_t *timer_ctx, unsigned fps);
 
 typedef struct {
